reorganize Test project and create sample Controller Test
