+++ WebStoreASP.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29025.244
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStoreTests", "Tests\WebStoreTests\WebStoreTests.csproj", "{E90448EB-64F2-44E4-ACFD-13A0DD0229DC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.Tests", "Tests\WebStore.Tests\WebStore.Tests.csproj", "{E90448EB-64F2-44E4-ACFD-13A0DD0229DC}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{C430DABF-5CB5-40B4-8866-E441DB3D4A88}"
 EndProject
@@ -27,7 +27,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.ServiceHosting", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.Clients", "Services\WebStore.Clients\WebStore.Clients.csproj", "{D5E9B3C3-A489-43B3-B008-1255BF04FCED}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebStore.Logger", "Common\WebStore.Logger\WebStore.Logger.csproj", "{C7F57F47-288A-4F99-8D25-6446882416C8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.Logger", "Common\WebStore.Logger\WebStore.Logger.csproj", "{C7F57F47-288A-4F99-8D25-6446882416C8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -71,6 +71,10 @@ Global
 		{C7F57F47-288A-4F99-8D25-6446882416C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C7F57F47-288A-4F99-8D25-6446882416C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C7F57F47-288A-4F99-8D25-6446882416C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4372EF7-8277-4004-9F09-07354356842F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4372EF7-8277-4004-9F09-07354356842F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4372EF7-8277-4004-9F09-07354356842F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4372EF7-8277-4004-9F09-07354356842F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
